requirements.txt
flask
flask-socketio
eventlet
mysql-connector-python
